requirements.txt
environs==11.0.0
iniconfig==2.0.0
marshmallow==3.23.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.3
python-dotenv==1.0.1
